+++ src/libming/blocks/outputblock.c
@@ -158,7 +158,7 @@ newSWFRemoveObject2Block(int depth)
 SWFOutputBlock
 newSWFFrameLabelBlock(const char *string)
 {
-	SWFOutput out = newSizedSWFOutput(strlen(string)+1);
+	SWFOutput out = newSizedSWFOutput((int) (strlen(string)+1));
 	SWFOutput_writeString(out, (byte*)string);
 
 	return newSWFOutputBlock(out, SWF_FRAMELABEL);
@@ -168,7 +168,7 @@ newSWFFrameLabelBlock(const char *string)
 SWFOutputBlock
 newSWFNamedAnchorBlock(const char *string)
 {
-	SWFOutput out = newSizedSWFOutput(strlen(string)+2);
+	SWFOutput out = newSizedSWFOutput((int) (strlen(string)+2));
 	SWFOutput_writeString(out, (byte*)string);
 	SWFOutput_writeUInt8(out, 1);
