+++ src/libming/shape_util.c
@@ -65,7 +65,7 @@ float SWFCharacter_getHeight(SWFCharacter character)
 void SWFShape_setLine(SWFShape shape, unsigned short width,
 					byte r, byte g, byte b, byte a)
 {
-	SWFShape_setLineStyle_internal(shape, width * Ming_scale, r, g, b, a);
+	SWFShape_setLineStyle_internal(shape, (unsigned short) (width * Ming_scale), r, g, b, a);
 }
 
 /*
@@ -112,7 +112,7 @@ void SWFShape_setLine2(SWFShape shape, unsigned short width,
                        byte r, byte g, byte b, byte a,
                        int flags, float miterLimit)
 {
-	SWFShape_setLineStyle2_internal(shape, width * Ming_scale,
+	SWFShape_setLineStyle2_internal(shape, (unsigned short) (width * Ming_scale),
 		r, g, b, a, flags, miterLimit);
 }
 
@@ -162,7 +162,7 @@ void SWFShape_setLine2Filled(SWFShape shape, unsigned short width,
                              SWFFillStyle fill,
                              int flags, float miterLimit)
 {
-	SWFShape_setLineStyle2filled_internal(shape, width * Ming_scale,
+	SWFShape_setLineStyle2filled_internal(shape, (unsigned short) (width * Ming_scale),
 		fill, flags, miterLimit);
 }
